Refactor HmdaApi
